+++ src/main/java/com/jinju/jinjuwiki/domain/auth/dto/request/EmailVerificationSendRequest.java
@@ -1,14 +1,11 @@
-package com.jinju.jinjuwiki.domain.auth.dto;
+package com.jinju.jinjuwiki.domain.auth.dto.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 
-public record LoginRequest(
+public record EmailVerificationSendRequest(
         @NotBlank(message = "이메일은 비어 있을 수 없습니다.")
         @Email(message = "올바른 이메일 형식이어야 합니다.")
-        String email,
-
-        @NotBlank(message = "비밀번호는 비어 있을 수 없습니다.")
-        String password
+        String email
 ) {
 }
