+++ src/main/java/com/jinju/jinjuwiki/domain/auth/dto/response/EmailVerificationVerifyResponse.java
@@ -1,4 +1,4 @@
-package com.jinju.jinjuwiki.domain.auth.dto;
+package com.jinju.jinjuwiki.domain.auth.dto.response;
 
 import java.time.LocalDateTime;
