+++ src/main/java/com/jinju/jinjuwiki/domain/auth/dto/response/LoginResponse.java
@@ -1,4 +1,4 @@
-package com.jinju.jinjuwiki.domain.auth.dto;
+package com.jinju.jinjuwiki.domain.auth.dto.response;
 
 public record LoginResponse(
         String accessToken,
