+++ src/main/java/com/jinju/jinjuwiki/domain/auth/dto/response/SignupResponse.java
@@ -1,4 +1,4 @@
-package com.jinju.jinjuwiki.domain.auth.dto;
+package com.jinju.jinjuwiki.domain.auth.dto.response;
 
 public record SignupResponse(
         Long userId,
