+++ src/main/java/com/jinju/jinjuwiki/domain/auth/service/AuthService.java
@@ -1,13 +1,13 @@
 package com.jinju.jinjuwiki.domain.auth.service;
 
-import com.jinju.jinjuwiki.domain.auth.dto.LoginRequest;
-import com.jinju.jinjuwiki.domain.auth.dto.LoginResponse;
-import com.jinju.jinjuwiki.domain.auth.dto.EmailVerificationSendRequest;
-import com.jinju.jinjuwiki.domain.auth.dto.EmailVerificationSendResponse;
-import com.jinju.jinjuwiki.domain.auth.dto.EmailVerificationVerifyRequest;
-import com.jinju.jinjuwiki.domain.auth.dto.EmailVerificationVerifyResponse;
-import com.jinju.jinjuwiki.domain.auth.dto.SignupRequest;
-import com.jinju.jinjuwiki.domain.auth.dto.SignupResponse;
+import com.jinju.jinjuwiki.domain.auth.dto.request.LoginRequest;
+import com.jinju.jinjuwiki.domain.auth.dto.response.LoginResponse;
+import com.jinju.jinjuwiki.domain.auth.dto.request.EmailVerificationSendRequest;
+import com.jinju.jinjuwiki.domain.auth.dto.response.EmailVerificationSendResponse;
+import com.jinju.jinjuwiki.domain.auth.dto.request.EmailVerificationVerifyRequest;
+import com.jinju.jinjuwiki.domain.auth.dto.response.EmailVerificationVerifyResponse;
+import com.jinju.jinjuwiki.domain.auth.dto.request.SignupRequest;
+import com.jinju.jinjuwiki.domain.auth.dto.response.SignupResponse;
 
 public interface AuthService {
